backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
python-dotenv
